scheduler/requirements.txt
fastapi==0.65.1
uvicorn==0.14.0
pydantic==1.8.2
python-multipart==0.0.5
passlib==1.7.4
async-exit-stack==1.0.1
async-generator==1.10
python-crontab==2.5.1
aiofiles==0.7.0
itsdangerous==2.0.1
tortoise-orm==0.17.4
aerich==0.5.3
# git+https://github.com/ShAlireza/BO/tree/master/packages/bo_shared
pika==1.2.0
asyncpg==0.23.0
